# ApiDockerTecnimotors/ApiDockerTecnimotors/obj/Debug/net8.0/EndpointInfo/ApiDockerTecnimotors.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ApiDockerTecnimotors",
    "version": "1.0"
  },
  "paths": {
...
    "/api/MaestroClasificado/ListadoGeneralCategoria/{motocicleta}": {
      "get": {
        "tags": [
          "MaestroClasificado"
        ],
        "parameters": [
          {
            "name": "motocicleta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/ListadoGeneralVehiculos": {
      "post": {
        "tags": [
          "MaestroClasificado"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralVehiculo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralVehiculo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralVehiculo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/ListadoModeloVehiculo": {
      "post": {
        "tags": [
          "MaestroClasificado"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralModelo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralModelo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralModelo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/ListarMarcaVehiculo": {
      "post": {
        "tags": [
          "MaestroClasificado"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralMarca"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralMarca"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GeneralMarca"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/DetalleVehiculo/{Id}": {
      "get": {
        "tags": [
          "MaestroClasificado"
        ],
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/MaestroClasificado/GetBanner1/{ruta}": {
      "get": {
        "tags": [
          "MaestroClasificado"
        ],
        "parameters": [
          {
            "name": "ruta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/GetBanner2": {
      "get": {
        "tags": [
          "MaestroClasificado"
        ],
        "parameters": [
          {
            "name": "ruta",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/GetBanner3": {
      "post": {
        "tags": [
          "MaestroClasificado"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Ruta": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Ruta": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MaestroClasificado/GetBanner4": {
      "post": {
        "tags": [
          "MaestroClasificado"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trruta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Trruta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Trruta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GeneralMarca": {
        "type": "object",
        "properties": {
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "medida": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GeneralModelo": {
        "type": "object",
        "properties": {
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GeneralVehiculo": {
        "type": "object",
        "properties": {
          "categoria": {
            "type": "string",
            "nullable": true
          },
          "medida": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Trruta": {
        "type": "object",
        "properties": {
          "ruta": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}